requirements.txt
flask
langchain-groq
langchain-core
python-dotenv
langchain
